Initial attempt to remove the fmtutil.cnf scriptlets
This contains the basic work to remove the scriptlets which modify
fmtutil.cnf from the package.

This takes the fmtutil.cnf file from the distribution and splits it into
individual files. Those files are added to the relevant packages. A
small script is installed which puts together the individual files into
one fmtutil.cnf file. Two file trigger scriptlets call that when
necessary.

Most of the diff is renumbering 350+ source files to make room for
one at the beginning of the list.